+++ test/bin/cache_speed_test.ml
@@ -1,3 +1,0 @@
-open! Core
-
-let () = Command.run Test_async_smtp.Cache_speed_test.command